Moved eval to plugin
