+++ src/main/index.ts
@@ -23,6 +23,7 @@ import enableBigIntEssentials from './plugin/bigint-essentials';
 import enableDateEssentials from './plugin/date-essentials';
 import enableNumberEssentials from './plugin/number-essentials';
 import enableObjectEssentials from './plugin/object-essentials';
+import enableObjectEval from './plugin/object-eval';
 import enableRecordEssentials from './plugin/record-essentials';
 import enableSetEssentials from './plugin/set-essentials';
 import enableStringEssentials from './plugin/string-essentials';
@@ -43,6 +44,7 @@ enableBigIntEssentials(BigIntShape);
 enableDateEssentials(DateShape);
 enableNumberEssentials(NumberShape);
 enableObjectEssentials(ObjectShape);
+enableObjectEval(ObjectShape);
 enableRecordEssentials(RecordShape);
 enableSetEssentials(SetShape);
 enableStringEssentials(StringShape);
