+++ src/main/plugin/shape-utils.ts
@@ -1,0 +1,13 @@
+/**
+ * `true` if runtime code evaluation is supported.
+ *
+ * @alpha
+ */
+export const isEvalSupported = (() => {
+  try {
+    new Function('');
+    return true;
+  } catch {
+    return false;
+  }
+})();
